feat:  utils.ts with func from delivery AM
